requirements.txt
aniso8601==9.0.1
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
cymem==2.0.5
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dnspython==1.16.0
entrypoints==0.3
Flask==2.0.1
Flask-Cors==3.0.10
flask-restplus==0.13.0
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lime==0.2.0.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.3
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.1.0/pt_core_news_sm-3.1.0-py3-none-any.whl
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
terminado==0.12.1
testpath==0.5.0
thinc==8.0.10
threadpoolctl==2.2.0
tifffile==2021.8.30
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
tweepy==3.10.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
